requirements.txt
flask==1.1.0
gunicorn==20.0.4
requests==2.25.1
pandas==1.3.5
bokeh==2.4.2
scikit-learn==1.0.1
python-dotenv